Handle duplicate UE IP and TEIDs across multiple sessions
This patch makes sure that a PDIs don't conflict for the PDIs for the
same NWI from other sessions based on UE IPs and TEIDs. This is not
yet a complete implementation of the spec in this respect b/c other
IEs in PDI also need to be taken into account.

Previously, adding many duplicate UE IPs could cause UPG to
crash (about 255) due to FIB entry src count limit, and, in addition
to that, deleting sessions with duplicate UEs could lead to an
assertion failure due to a bad session reference.